requirements-lite.txt
# Dependências base (leves, sempre necessárias)
pdfplumber>=0.11.0
streamlit>=1.30.0
markdown>=3.6.0
beautifulsoup4>=4.12.0
python-Levenshtein>=0.25.0
markitdown>=0.1.0
pymupdf4llm>=0.0.10
